requirements.txt
absl-py==0.13.0
adal==1.2.7
aiohttp==3.7.4.post0
alabaster==0.7.12
antlr4-python3-runtime==4.8
apex==0.1
appdirs==1.4.4
argcomplete==2.1.1
argon2-cffi==20.1.0
ascii-graph==1.5.1
ase==3.22.1
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
azure-common==1.1.28
azure-core==1.26.3
azure-graphrbac==0.61.1
azure-identity==1.12.0
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azure-storage-blob==12.13.0
azureml-core==1.49.0
azureml-mlflow==1.49.0
babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
beautifulsoup4==4.9.3
bitarray==2.7.3
bleach==3.3.0
blis==0.7.4
boto3==1.17.96
boto==2.49.0
botocore==1.20.96
brotlipy==0.7.0
bz2file==0.98
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.2.1
codecov==2.1.11
colorama==0.4.4
conda-build==3.21.4
conda-package-handling==1.7.3
conda==4.10.1
contextlib2==21.6.0
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
cython==0.29.30
databricks-cli==0.17.4
dataproperty==0.50.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.6
django==3.2
dllogger==0.1.0
docker-pycreds==0.4.0
docker==6.0.1
docutils==0.16
einops==0.4.1
entrypoints==0.3
fairseq==0.12.2
filelock==3.0.12
flake8==3.7.9
flask==2.0.1
flatbuffers==2.0
fonttools==4.39.0
fsspec==2023.3.0
future==0.18.2
gitdb==4.0.10
gitpython==3.1.31
glob2==0.7
google-api-core==1.30.0
google-auth-oauthlib==0.4.4
google-auth==1.30.2
google-cloud-core==1.7.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
graphsurgeon==0.4.5
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.2.1
html2text==2020.1.16
httptools==0.2.0
humanfriendly==10.0
hydra-core==1.0.7
hypothesis==4.50.8
idna==2.10
imageio==2.19.3
imagesize==1.2.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==5.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==5.5.5
ipython-genutils==0.2.0
ipython==7.24.1
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.8.0
jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonpickle==2.2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab==2.3.1
jupytext==1.11.3
kiwisolver==1.3.1
knack==0.10.1
libarchive-c==3.1
librosa==0.8.0
llvmlite==0.35.0
lmdb==1.3.0
lxml==4.9.2
mako==1.1.4
markdown-it-py==1.1.0
markdown==3.3.4
markupsafe==2.0.1
maskrcnn-benchmark==0.1
matplotlib-inline==0.1.2
matplotlib==3.5.2
mbstrdecoder==1.0.1
mccabe==0.6.1
mdit-py-plugins==0.2.8
mistune==0.8.4
mlflow-skinny==2.2.1
mlperf-compliance==0.0.10
mock==4.0.3
monai==0.9.0
mpi4py==3.1.3
mpmath==1.3.0
msal-extensions==1.0.0
msal==1.21.0
msgfy==0.1.0
msrest==0.7.1
msrestazure==0.6.4
multidict==5.1.0
multiprocess==0.70.14
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.1
networkx==2.8
nltk==3.6.2
notebook==6.2.0
numba==0.52.0
numpy==1.22.3
nvidia-dali-cuda110==1.2.0
nvidia-dlprof-pytorch-nvtx==1.2.0
nvidia-dlprofviewer==1.2.0
nvidia-ml-py==11.515.48
oauthlib==3.1.1
omegaconf==2.0.6
onnx==1.8.0
onnxruntime==1.8.0
opencv-python-headless==4.7.0.72
opencv-python==4.6.0.66
packaging==20.9
pandas==1.4.2
pandocfilters==1.4.3
paramiko==2.12.0
parso==0.8.2
pathos==0.2.9
pathspec==0.11.0
pathtools==0.1.2
pathvalidate==2.4.1
pathy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pillow-simd==7.0.0.post3
pillow==9.4.0
pip==21.1.2
pkginfo==1.7.0
pluggy==0.13.1
polygraphy==0.30.2
pooch==1.4.0
portalocker==2.7.0
positional-encodings==6.0.1
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.5
prettytable==2.1.0
progressbar==2.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1-modules==0.2.7
pyasn1==0.4.8
pybind11==2.6.2
pycocotools==2.0+nv0.5.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycuda==2020.1
pydantic==1.7.3
pydeprecate==0.3.1
pydot==1.4.2
pyflakes==2.1.1
pygments==2.9.0
pyjwt==2.6.0
pynacl==1.5.0
pynvml==8.0.4
pyopenssl==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyscf==2.0.1
pysocks==1.7.1
pytablewriter==0.47.0
pytest-cov==2.12.1
pytest-pythonpath==0.7.3
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-hostlist==1.21
python-nvd3==0.15.0
python-slugify==5.0.2
pytools==2021.2.7
pytorch-lightning==1.5.0
pytorch-quantization==2.1.0
pytorch-transformers==1.1.0
pytz==2021.1
pyu2f==0.1.5
pywavelets==1.1.1
pyyaml==5.4.1
pyzmq==22.1.0
ratelimit==2.2.1
rdflib==6.2.0
rdkit==2022.3.5
regex==2021.4.4
requests-oauthlib==1.3.0
requests==2.28.0
resampy==0.2.2
revtok==0.0.3
rsa==4.7.2
ruamel-yaml-conda==0.15.80
s3transfer==0.4.2
sacrebleu==2.3.1
sacremoses==0.0.35
schnetpack==1.0.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.0
secretstorage==3.3.3
send2trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==49.6.0.post20210108
shellingham==1.4.0
six==1.16.0
smart-open==2.2.1
smmap==5.0.0
snowballstemmer==2.1.0
soundfile==0.10.3.post1
soupsieve==2.0.1
sox==1.4.1
spacy-legacy==3.0.6
spacy==3.0.6
sphinx-glpi-theme==0.3
sphinx-rtd-theme==0.5.2
sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
srsly==2.4.1
subword-nmt==0.3.3
supervisor==4.2.5
sympy==1.11.1
tabledata==1.1.4
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.5.0
tensorboardx==2.6
tensorrt==7.2.3.4
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.3
threadpoolctl==2.1.0
tifffile==2023.2.28
toml==0.10.2
torch-cluster==1.5.9
torch-ema==0.3
torch-geometric==2.0.3
torch-scatter==2.0.8
torch-sparse==0.6.10
torch==1.9.0+cu111
torchaudio==0.9.0
torchinfo==1.6.6
torchmetrics==0.11.3
torchtext==0.10.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.0.5
typepy==1.1.5
typer==0.3.2
typing-extensions==4.5.0
uff==0.6.9
unidecode==1.2.0
urllib3==1.26.14
uvicorn==0.14.0
uvloop==0.15.2
volumentations-3d==1.0.4
wandb==0.13.11
wasabi==0.8.2
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
websockets==9.1
werkzeug==2.0.1
wheel==0.36.2
whitenoise==5.2.0
wrapt==1.10.11
yacs==0.1.8
yarl==1.5.1
zipp==3.15.0
